+++ NEWS.md
@@ -1,7 +1,10 @@
 StMoMo v0.2.0.9000
 ----------------------------------------------------------------
-- Test new PC git 5
-
+* Add optional parameters kt.lookbacak and gc.lookback to 
+  functions forecast.fitStMoMo, simulate.fitStMoMo and 
+  simulate.bootStMoMo. These new parameter allow the forecast
+  and simulation of mortality models using only a limited history
+  of the period of the period and cohort indexes
 
 StMoMo v0.2.0
 ----------------------------------------------------------------